moved your teams to all teams page
