+++ client/src/pages/TeamsPage.js
@@ -1,19 +1,30 @@
 import { NavLink, useNavigate } from "react-router-dom";
 import { useLoaderData } from "react-router-dom";
 import { useAuth } from "../context/AuthContext";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
+import axios from "axios";
 
 const TeamsPage = () => {
   const { authedUser } = useAuth();
   const { allTeamsData } = useLoaderData();
   const navigate = useNavigate();
   const { teams } = allTeamsData.data;
+  const authedUserId = authedUser?.id;
+  const [userTeams, setUserTeams] = useState([]);
 
   useEffect(() => {
     if (!authedUser) {
       navigate("/login");
     }
   }, [authedUser]);
+
+  useEffect(() => {
+    const getUserTeams = async () => {
+      const response = await axios.get(`/api/users/${authedUserId}/teams`);
+      setUserTeams(response.data.teams);
+    };
+    getUserTeams();
+  }, [authedUserId]);
 
   return (
     <div className="flex flex-col">
@@ -24,6 +35,18 @@ const TeamsPage = () => {
             to={`/teams/${team.id}`}
             className="bg-white p-2.5 border-t-[0.5px] border-l-[0.5px] rounded-sm shadow-[0_0.3px_1px_rgba(0,0,0,0.2)] hover:bg-blue-200"
             key={`${team.name}-${index}`}
+          >
+            {team.name}
+          </NavLink>
+        ))}
+      </div>
+      <div className="flex flex-col">
+        <h1 className="text-2xl font-bold">Your Teams</h1>
+        {userTeams.map((team) => (
+          <NavLink
+            to={`/teams/${team.id}`}
+            key={team.id}
+            className="bg-white p-2.5 border-t-[0.5px] border-l-[0.5px] rounded-sm shadow-[0_0.3px_1px_rgba(0,0,0,0.2)] hover:bg-blue-200"
           >
             {team.name}
           </NavLink>
